+++ Source/Examples/ExampleLibrary/Series/HistogramSeriesExamples.cs
@@ -90,6 +90,12 @@ namespace ExampleLibrary
             return CreateDisconnectedBins();
         }
 
+        [Example("Normal Distribution Three Colors")]
+        public static PlotModel NormalDistribution()
+        {
+            return CreateNormalDistribution();
+        }
+
         public static PlotModel CreateExponentialDistribution(double mean = 1, int n = 10000)
         {
             var model = new PlotModel { Title = "Exponential Distribution", Subtitle = "Uniformly distributed bins (" + n + " samples)" };
@@ -128,6 +134,43 @@ namespace ExampleLibrary
             return model;
         }
 
+        public static PlotModel CreateNormalDistribution(double mean = 0, double std = 1, int n = 1000000)
+        {
+            var model = new PlotModel { Title = $"Normal Distribution (μ={mean}, σ={std})", Subtitle = "95% of the distribution (" + n + " samples)" };
+            model.Axes.Add(new LinearAxis { Position = AxisPosition.Left, Title = "Frequency" });
+            model.Axes.Add(new LinearAxis { Position = AxisPosition.Bottom, Title = "x" });
+
+            Random rnd = new Random();
+
+            HistogramSeries chs = new HistogramSeries();
+            var binningOptions = new BinningOptions(BinningOutlierMode.CountOutliers, BinningIntervalType.InclusiveLowerBound, BinningExtremeValueMode.ExcludeExtremeValues);
+            var binBreaks = HistogramHelpers.CreateUniformBins(-std * 4, std * 4, 100);
+            chs.Items.AddRange(HistogramHelpers.Collect(SampleNormal(rnd, mean, std, n), binBreaks, binningOptions));
+            chs.StrokeThickness = 1;
+
+            double LimitHi = mean + 1.96 * std;
+            double LimitLo = mean - 1.96 * std;
+            OxyColor ColorHi = OxyColors.DarkRed;
+            OxyColor ColorLo = OxyColors.DarkRed;
+
+            chs.ColorMapping = (item) =>
+            {
+                if (item.RangeCenter > LimitHi)
+                {
+                    return ColorHi;
+                }
+                else if (item.RangeCenter < LimitLo)
+                {
+                    return ColorLo;
+                }
+                return chs.ActualFillColor;
+            };
+
+            model.Series.Add(chs);
+
+            return model;
+        }
+
         public static PlotModel CreateDisconnectedBins()
         {
             var model = new PlotModel { Title = "Disconnected Bins" };
@@ -154,6 +197,22 @@ namespace ExampleLibrary
         private static double SampleExp(Random rnd, double mean)
         {
             return Math.Log(1.0 - rnd.NextDouble()) / -mean;
+        }
+
+        private static IEnumerable<double> SampleNormal(Random rnd, double mean, double std, int count)
+        {
+            for (int i = 0; i < count; i++)
+            {
+                yield return SampleNormal(rnd, mean, std);
+            }
+        }
+
+        private static double SampleNormal(Random rnd, double mean, double std)
+        {
+            // http://en.wikipedia.org/wiki/Box%E2%80%93Muller_transform
+            var u1 = rnd.NextDouble();
+            var u2 = rnd.NextDouble();
+            return Math.Sqrt(-2 * Math.Log(u1)) * Math.Cos(2 * Math.PI * u2);
         }
     }
 }